fix: re-render of pages section
