+++ app/client/src/pages/Editor/IDE/EditorPane/index.tsx
@@ -4,35 +4,19 @@ import type { RouteComponentProps } from "react-router";
 import { Switch } from "react-router";
 import { useSelector } from "react-redux";
 
-import Pages from "pages/Editor/Explorer/Pages";
 import { SentryRoute } from "@appsmith/AppRouter";
 import { ADD_PATH } from "constants/routes";
 import EditorPaneSegments from "./EditorPaneSegments";
 import GlobalAdd from "./GlobalAdd";
 import { useEditorPaneWidth } from "../hooks";
-import {
-  getIsSideBySideEnabled,
-  getPagesActiveStatus,
-} from "selectors/ideSelectors";
+import { getPagesActiveStatus } from "selectors/ideSelectors";
 import EntityProperties from "pages/Editor/Explorer/Entity/EntityProperties";
-import { PagesSection } from "./PagesSection";
 import { MinimalSegment } from "./MinimalSegment";
+import { Pages } from "./components/Pages";
 
 const EditorPane = ({ match: { path } }: RouteComponentProps) => {
   const width = useEditorPaneWidth();
   const pagesActive = useSelector(getPagesActiveStatus);
-  const isSideBySideEnabled = useSelector(getIsSideBySideEnabled);
-
-  const PagesRender = () => {
-    if (!isSideBySideEnabled) {
-      return <Pages />;
-      /* divider is inside the Pages component */
-    } else if (isSideBySideEnabled && pagesActive) {
-      return <PagesSection />;
-    } else {
-      return null;
-    }
-  };
 
   return (
     <Flex
@@ -46,7 +30,7 @@ const EditorPane = ({ match: { path } }: RouteComponentProps) => {
       {/** Entity Properties component is needed to render
         the Bindings popover in the context menu. Will be removed eventually **/}
       <EntityProperties />
-      <PagesRender />
+      <Pages />
 
       {pagesActive ? (
         <MinimalSegment />
